refactor: resolve NG short-link base URL in container
Move the default Neuroglancer base URL resolution out of NGLinkDialog into a
useDefaultNeuroglancerBaseUrl hook used by the NGLinks container, which passes
the result down as a prop. Keeps the dialog a pure presentational component (no
direct context access), so it renders without the viewers/preferences provider
tree and its unit tests pass without extra wrappers. No behavior change.
